fix(reload): same-file reload wiring, TF invalidation, visible-topic timeline (#166)
Three reload-path bugs plus the compile-time guard that would have caught
the first one:

- FileLoader: re-target the reload ingest at the staged engine/store pair
  (PR #135's merge reverted PR #124's wiring, so a same-file reload
  re-pushed every message into the live store — ObjectStore monotonicity
  warning flood — then swapped from the empty staging area, wiping the
  dataset's topics and emptying the curve tree until restart).
- TransformService: add invalidateDataset() and call it on the in-place
  swap, on tombstoned fanout-replace datasets, and on dataset removal —
  the populated-guard otherwise skips the re-ingest and 3D views keep the
  previous load's transforms. Fanout entries now TF-ingest their own live
  dataset ids (the staged scratch id could alias an unrelated dataset).
- AppSession::seedPlaybackFromSession: recompute the playback range from
  the catalog-visible topics (per-topic granularity) instead of all engine
  datasets with a monotonic union — removed/trashed data no longer
  stretches the timeline, the range can shrink, and the first-load snap
  re-arms when the catalog empties. MainWindow re-seeds at removal sites
  (guarded while a streaming dataset owns the slider).
- pj_app: enable PJ_WARNING_FLAGS (-Werror) and fix the 18 outstanding
  warnings; an unused-variable error would have caught the dropped reload
  wiring at compile time. FileLoader no longer links MainWindow (file
  picker injected via setFilePicker), making it testable headlessly.

Tests (all written first and watched fail): file_loader_test drives the
real plugin pipeline through the SDK's mock_file_source_plugin (same-file
reload + reload-one-of-many), and three new app_session_test cases pin
the timeline-range semantics. Verified against a real MCAP + the real
loader plugin: zero monotonicity warnings on reload, stable catalog, and
TF re-ingested (82k transforms) instead of skipped.

+++ pj_app/src/ui/Scene2DConfigPanel.cpp
@@ -19,7 +19,7 @@ namespace {
 // opaque qint64. The dock's topic id IS that opaque key.
 [[nodiscard]] ObjectTopicId toTopicId(qint64 id) {
   ObjectTopicId tid;
-  tid.id = static_cast<int64_t>(id);
+  tid.id = static_cast<uint32_t>(id);
   return tid;
 }
 }  // namespace
